fix: Hourly Chart Reflects 24 Consecutive Hours
Previously the temperature by the hour chart would display the first
reading of the hour. This created introduced some styling difficulties
that were caused when observing the chart at 12:00 AM or at 11:00 PM.
During these noted times users would observe either no data, or too much
data. In the case of no data it appeared that the chart was not
functioning. In the case of too much data, the chart would not display
the h-axis labels correctly.

This change presented an issue where getting the first reading of the
hour would result in only 1 hour of a 24 hour day being displayed. For
Ex. If current time is 3/18/2016 7:15 PM Displaying 3/17 7:00 PM and 3/18 7:00 PM is not possible in the scripts
current implementation. This is because the 7:00 hour can only be
displayed once. As a result running this script would display 3/17
7:16 PM and 3/18 7:00 PM. This behavior was corrected by excluding the
first reading.